[HUDI-9564] Inroduce BufferRecordMerger API in FileGroupRecordBuffer
update

fix comments

update

cosmetic changes
