+++ hudi-common/src/main/java/org/apache/hudi/common/table/read/FileGroupRecordBuffer.java
@@ -23,26 +23,18 @@ import org.apache.hudi.avro.AvroSchemaCache;
 import org.apache.hudi.common.config.RecordMergeMode;
 import org.apache.hudi.common.config.TypedProperties;
 import org.apache.hudi.common.engine.HoodieReaderContext;
-import org.apache.hudi.common.model.DefaultHoodieRecordPayload;
 import org.apache.hudi.common.model.DeleteRecord;
-import org.apache.hudi.common.model.HoodieAvroRecord;
-import org.apache.hudi.common.model.HoodieKey;
 import org.apache.hudi.common.model.HoodieOperation;
 import org.apache.hudi.common.model.HoodiePayloadProps;
-import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.model.HoodieRecordMerger;
-import org.apache.hudi.common.model.OverwriteWithLatestAvroPayload;
 import org.apache.hudi.common.table.HoodieTableConfig;
 import org.apache.hudi.common.table.HoodieTableMetaClient;
 import org.apache.hudi.common.table.log.KeySpec;
 import org.apache.hudi.common.table.log.block.HoodieDataBlock;
 import org.apache.hudi.common.util.DefaultSizeEstimator;
 import org.apache.hudi.common.util.FileIOUtils;
-import org.apache.hudi.common.util.HoodieRecordUtils;
 import org.apache.hudi.common.util.InternalSchemaCache;
 import org.apache.hudi.common.util.Option;
-import org.apache.hudi.common.util.ReflectionUtils;
-import org.apache.hudi.common.util.ValidationUtils;
 import org.apache.hudi.common.util.collection.ClosableIterator;
 import org.apache.hudi.common.util.collection.CloseableMappingIterator;
 import org.apache.hudi.common.util.collection.ExternalSpillableMap;
@@ -53,15 +45,12 @@ import org.apache.hudi.internal.schema.action.InternalSchemaMerger;
 import org.apache.hudi.internal.schema.convert.AvroInternalSchemaConverter;
 
 import org.apache.avro.Schema;
-import org.apache.avro.generic.GenericRecord;
-import org.apache.avro.generic.IndexedRecord;
 
 import java.io.IOException;
 import java.io.Serializable;
 import java.util.Iterator;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
 import java.util.function.Function;
 
 import static org.apache.hudi.common.config.HoodieCommonConfig.DISK_MAP_BITCASK_COMPRESSION_ENABLED;
@@ -70,7 +59,6 @@ import static org.apache.hudi.common.config.HoodieMemoryConfig.MAX_MEMORY_FOR_ME
 import static org.apache.hudi.common.config.HoodieMemoryConfig.SPILLABLE_MAP_BASE_PATH;
 import static org.apache.hudi.common.model.HoodieRecord.DEFAULT_ORDERING_VALUE;
 import static org.apache.hudi.common.model.HoodieRecord.HOODIE_IS_DELETED_FIELD;
-import static org.apache.hudi.common.model.HoodieRecord.OPERATION_METADATA_FIELD;
 import static org.apache.hudi.common.model.HoodieRecordMerger.PAYLOAD_BASED_MERGE_STRATEGY_UUID;
 import static org.apache.hudi.common.table.log.block.HoodieLogBlock.HeaderMetadataType.INSTANT_TIME;
 
@@ -93,6 +81,7 @@ public abstract class FileGroupRecordBuffer<T> implements HoodieFileGroupRecordB
   protected boolean enablePartialMerging = false;
   protected InternalSchema internalSchema;
   protected HoodieTableMetaClient hoodieTableMetaClient;
+  protected BufferedRecordMerger<T> bufferedRecordMerger;
   private long totalLogRecords = 0;
 
   protected FileGroupRecordBuffer(HoodieReaderContext<T> readerContext,
@@ -141,6 +130,8 @@ public abstract class FileGroupRecordBuffer<T> implements HoodieFileGroupRecordB
         readerContext.getSchemaHandler().getCustomDeleteMarkerKeyValue().isPresent();
     this.shouldCheckBuiltInDeleteMarker =
         readerContext.getSchemaHandler().hasBuiltInDelete();
+    this.bufferedRecordMerger = BufferedRecordMergerFactory.create(
+        readerContext, recordMergeMode, enablePartialMerging, recordMerger, orderingFieldName, payloadClass, readerSchema, props);
   }
 
   /**
@@ -243,92 +234,7 @@ public abstract class FileGroupRecordBuffer<T> implements HoodieFileGroupRecordB
   protected Option<BufferedRecord<T>> doProcessNextDataRecord(BufferedRecord<T> newRecord, BufferedRecord<T> existingRecord)
       throws IOException {
     totalLogRecords++;
-    if (existingRecord != null) {
-      if (enablePartialMerging) {
-        // TODO(HUDI-7843): decouple the merging logic from the merger
-        //  and use the record merge mode to control how to merge partial updates
-        // Merge and store the combined record
-        Option<Pair<HoodieRecord, Schema>> combinedRecordAndSchemaOpt = recordMerger.get().partialMerge(
-            readerContext.constructHoodieRecord(existingRecord),
-            readerContext.getSchemaFromBufferRecord(existingRecord),
-            readerContext.constructHoodieRecord(newRecord),
-            readerContext.getSchemaFromBufferRecord(newRecord),
-            readerSchema,
-            props);
-        if (!combinedRecordAndSchemaOpt.isPresent()) {
-          return Option.empty();
-        }
-        Pair<HoodieRecord, Schema> combinedRecordAndSchema = combinedRecordAndSchemaOpt.get();
-        HoodieRecord<T> combinedRecord = combinedRecordAndSchema.getLeft();
-
-        // If pre-combine returns existing record, no need to update it
-        if (combinedRecord.getData() != existingRecord.getRecord()) {
-          return Option.of(BufferedRecord.forRecordWithContext(combinedRecord, combinedRecordAndSchema.getRight(), readerContext, props));
-        }
-        return Option.empty();
-      } else {
-        switch (recordMergeMode) {
-          case COMMIT_TIME_ORDERING:
-            return Option.of(newRecord);
-          case EVENT_TIME_ORDERING:
-            if (shouldKeepNewerRecord(existingRecord, newRecord)) {
-              return Option.of(newRecord);
-            }
-            return Option.empty();
-          case CUSTOM:
-          default:
-            if (existingRecord.isDelete() || newRecord.isDelete()) {
-              if (shouldKeepNewerRecord(existingRecord, newRecord)) {
-                // IMPORTANT:
-                // this is needed when the fallback HoodieAvroRecordMerger got used, the merger would
-                // return Option.empty when the old payload data is empty(a delete) and ignores its ordering value directly.
-                return Option.of(newRecord);
-              } else {
-                return Option.empty();
-              }
-            }
-            // Merge and store the combined record
-            if (payloadClass.isPresent()) {
-              Option<Pair<HoodieRecord, Schema>> combinedRecordAndSchemaOpt = getMergedRecord(existingRecord, newRecord);
-              if (combinedRecordAndSchemaOpt.isPresent()) {
-                T combinedRecordData = readerContext.convertAvroRecord((IndexedRecord) combinedRecordAndSchemaOpt.get().getLeft().getData());
-                // If pre-combine does not return existing record, update it
-                if (combinedRecordData != existingRecord.getRecord()) {
-                  Pair<HoodieRecord, Schema> combinedRecordAndSchema = combinedRecordAndSchemaOpt.get();
-                  return Option.of(BufferedRecord.forRecordWithContext(combinedRecordData, combinedRecordAndSchema.getRight(), readerContext, orderingFieldName, false));
-                }
-              }
-              return Option.empty();
-            } else {
-              Option<Pair<HoodieRecord, Schema>> combinedRecordAndSchemaOpt = recordMerger.get().merge(
-                  readerContext.constructHoodieRecord(existingRecord),
-                  readerContext.getSchemaFromBufferRecord(existingRecord),
-                  readerContext.constructHoodieRecord(newRecord),
-                  readerContext.getSchemaFromBufferRecord(newRecord),
-                  props);
-
-              if (!combinedRecordAndSchemaOpt.isPresent()) {
-                return Option.empty();
-              }
-
-              Pair<HoodieRecord, Schema> combinedRecordAndSchema = combinedRecordAndSchemaOpt.get();
-              HoodieRecord<T> combinedRecord = combinedRecordAndSchema.getLeft();
-
-              // If pre-combine returns existing record, no need to update it
-              if (combinedRecord.getData() != existingRecord.getRecord()) {
-                return Option.of(BufferedRecord.forRecordWithContext(combinedRecord, combinedRecordAndSchema.getRight(), readerContext, props));
-              }
-              return Option.empty();
-            }
-        }
-      }
-    } else {
-      // Put the record as is
-      // NOTE: Record have to be cloned here to make sure if it holds low-level engine-specific
-      //       payload pointing into a shared, mutable (underlying) buffer we get a clean copy of
-      //       it since these records will be put into records(Map).
-      return Option.of(newRecord);
-    }
+    return bufferedRecordMerger.deltaMerge(newRecord, existingRecord);
   }
 
   /**
@@ -341,31 +247,7 @@ public abstract class FileGroupRecordBuffer<T> implements HoodieFileGroupRecordB
    */
   protected Option<DeleteRecord> doProcessNextDeletedRecord(DeleteRecord deleteRecord, BufferedRecord<T> existingRecord) {
     totalLogRecords++;
-    if (existingRecord != null) {
-      switch (recordMergeMode) {
-        case COMMIT_TIME_ORDERING:
-          return Option.of(deleteRecord);
-        case EVENT_TIME_ORDERING:
-        case CUSTOM:
-        default:
-          if (existingRecord.isCommitTimeOrderingDelete()) {
-            return Option.empty();
-          }
-          Comparable existingOrderingVal = existingRecord.getOrderingValue();
-          Comparable deleteOrderingVal = deleteRecord.getOrderingValue();
-          // Checks the ordering value does not equal to 0
-          // because we use 0 as the default value which means natural order
-          boolean chooseExisting = !deleteOrderingVal.equals(0)
-              && ReflectionUtils.isSameClass(existingOrderingVal, deleteOrderingVal)
-              && existingOrderingVal.compareTo(deleteOrderingVal) > 0;
-          if (chooseExisting) {
-            // The DELETE message is obsolete if the old message has greater orderingVal.
-            return Option.empty();
-          }
-      }
-    }
-    // Do delete.
-    return Option.of(deleteRecord);
+    return bufferedRecordMerger.deltaMerge(deleteRecord, existingRecord);
   }
 
   /**
@@ -425,138 +307,7 @@ public abstract class FileGroupRecordBuffer<T> implements HoodieFileGroupRecordB
    * @throws IOException
    */
   protected Pair<Boolean, T> merge(BufferedRecord<T> olderRecord, BufferedRecord<T> newerRecord) throws IOException {
-    if (enablePartialMerging) {
-      // TODO(HUDI-7843): decouple the merging logic from the merger
-      //  and use the record merge mode to control how to merge partial updates
-      Option<Pair<HoodieRecord, Schema>> mergedRecord = recordMerger.get().partialMerge(
-          readerContext.constructHoodieRecord(olderRecord), readerContext.getSchemaFromBufferRecord(olderRecord),
-          readerContext.constructHoodieRecord(newerRecord), readerContext.getSchemaFromBufferRecord(newerRecord),
-          readerSchema, props);
-
-      if (mergedRecord.isPresent()
-          && !mergedRecord.get().getLeft().isDelete(mergedRecord.get().getRight(), props)) {
-        HoodieRecord hoodieRecord = mergedRecord.get().getLeft();
-        if (!mergedRecord.get().getRight().equals(readerSchema)) {
-          T data = (T) hoodieRecord.rewriteRecordWithNewSchema(mergedRecord.get().getRight(), null, readerSchema).getData();
-          return Pair.of(false, data);
-        }
-        return Pair.of(false, (T) hoodieRecord.getData());
-      }
-      return Pair.of(true, null);
-    } else {
-      switch (recordMergeMode) {
-        case COMMIT_TIME_ORDERING:
-          return Pair.of(newerRecord.isDelete(), newerRecord.getRecord());
-        case EVENT_TIME_ORDERING:
-          if (newerRecord.isCommitTimeOrderingDelete()) {
-            return Pair.of(true, newerRecord.getRecord());
-          }
-          Comparable newOrderingValue = newerRecord.getOrderingValue();
-          Comparable oldOrderingValue = olderRecord.getOrderingValue();
-          if (!olderRecord.isCommitTimeOrderingDelete()
-              && oldOrderingValue.compareTo(newOrderingValue) > 0) {
-            return Pair.of(olderRecord.isDelete(), olderRecord.getRecord());
-          }
-          return Pair.of(newerRecord.isDelete(), newerRecord.getRecord());
-        case CUSTOM:
-        default:
-          if (payloadClass.isPresent()) {
-            if (olderRecord.isDelete() || newerRecord.isDelete()) {
-              if (shouldKeepNewerRecord(olderRecord, newerRecord)) {
-                // IMPORTANT:
-                // this is needed when the fallback HoodieAvroRecordMerger got used, the merger would
-                // return Option.empty when the new payload data is empty(a delete) and ignores its ordering value directly.
-                return Pair.of(newerRecord.isDelete(), newerRecord.getRecord());
-              } else {
-                return Pair.of(olderRecord.isDelete(), olderRecord.getRecord());
-              }
-            }
-            Option<Pair<HoodieRecord, Schema>> mergedRecord =
-                getMergedRecord(olderRecord, newerRecord);
-            if (mergedRecord.isPresent()
-                && !mergedRecord.get().getLeft().isDelete(mergedRecord.get().getRight(), props)) {
-              IndexedRecord indexedRecord;
-              if (!mergedRecord.get().getRight().equals(readerSchema)) {
-                indexedRecord = (IndexedRecord) mergedRecord.get().getLeft().rewriteRecordWithNewSchema(mergedRecord.get().getRight(), null, readerSchema).getData();
-              } else {
-                indexedRecord = (IndexedRecord) mergedRecord.get().getLeft().getData();
-              }
-              return Pair.of(false, readerContext.convertAvroRecord(indexedRecord));
-            }
-            return Pair.of(true, null);
-          } else {
-            if (olderRecord.isDelete() || newerRecord.isDelete()) {
-              if (shouldKeepNewerRecord(olderRecord, newerRecord)) {
-                // IMPORTANT:
-                // this is needed when the fallback HoodieAvroRecordMerger got used, the merger would
-                // return Option.empty when the new payload data is empty(a delete) and ignores its ordering value directly.
-                return Pair.of(newerRecord.isDelete(), newerRecord.getRecord());
-              } else {
-                return Pair.of(olderRecord.isDelete(), olderRecord.getRecord());
-              }
-            }
-            Option<Pair<HoodieRecord, Schema>> mergedRecord = recordMerger.get().merge(
-                readerContext.constructHoodieRecord(olderRecord), readerContext.getSchemaFromBufferRecord(olderRecord),
-                readerContext.constructHoodieRecord(newerRecord), readerContext.getSchemaFromBufferRecord(newerRecord), props);
-            if (mergedRecord.isPresent()
-                && !mergedRecord.get().getLeft().isDelete(mergedRecord.get().getRight(), props)) {
-              HoodieRecord hoodieRecord = mergedRecord.get().getLeft();
-              if (!mergedRecord.get().getRight().equals(readerSchema)) {
-                return Pair.of(false, (T) hoodieRecord.rewriteRecordWithNewSchema(mergedRecord.get().getRight(), null, readerSchema).getData());
-              }
-              return Pair.of(false, (T) hoodieRecord.getData());
-            }
-            return Pair.of(true, null);
-          }
-      }
-    }
-  }
-
-  /**
-   * Decides whether to keep the incoming record with ordering value comparison.
-   */
-  private boolean shouldKeepNewerRecord(BufferedRecord<T> oldRecord, BufferedRecord<T> newRecord) {
-    if (newRecord.isCommitTimeOrderingDelete()) {
-      // handle records coming from DELETE statements(the orderingVal is constant 0)
-      return true;
-    }
-    return newRecord.getOrderingValue().compareTo(oldRecord.getOrderingValue()) >= 0;
-  }
-
-  private Option<Pair<HoodieRecord, Schema>> getMergedRecord(BufferedRecord<T> olderRecord, BufferedRecord<T> newerRecord) throws IOException {
-    ValidationUtils.checkArgument(!Objects.equals(payloadClass, OverwriteWithLatestAvroPayload.class.getCanonicalName())
-        && !Objects.equals(payloadClass, DefaultHoodieRecordPayload.class.getCanonicalName()));
-    HoodieRecord oldHoodieRecord = constructHoodieAvroRecord(readerContext, olderRecord);
-    HoodieRecord newHoodieRecord = constructHoodieAvroRecord(readerContext, newerRecord);
-    Option<Pair<HoodieRecord, Schema>> mergedRecord = recordMerger.get().merge(
-        oldHoodieRecord, getSchemaForAvroPayloadMerge(oldHoodieRecord, olderRecord),
-        newHoodieRecord, getSchemaForAvroPayloadMerge(newHoodieRecord, newerRecord), props);
-    return mergedRecord;
-  }
-
-  /**
-   * Constructs a new {@link HoodieAvroRecord} for payload based merging
-   *
-   * @param readerContext reader context
-   * @param bufferedRecord buffered record
-   * @return A new instance of {@link HoodieRecord}.
-   */
-  private HoodieRecord constructHoodieAvroRecord(HoodieReaderContext<T> readerContext, BufferedRecord<T> bufferedRecord) {
-    GenericRecord record = null;
-    if (!bufferedRecord.isDelete()) {
-      Schema recordSchema = readerContext.getSchemaFromBufferRecord(bufferedRecord);
-      record = readerContext.convertToAvroRecord(bufferedRecord.getRecord(), recordSchema);
-    }
-    HoodieKey hoodieKey = new HoodieKey(bufferedRecord.getRecordKey(), null);
-    return new HoodieAvroRecord<>(hoodieKey,
-        HoodieRecordUtils.loadPayload(payloadClass.get(), record, bufferedRecord.getOrderingValue()), null);
-  }
-
-  private Schema getSchemaForAvroPayloadMerge(HoodieRecord record, BufferedRecord<T> bufferedRecord) throws IOException {
-    if (record.isDelete(readerSchema, props)) {
-      return readerSchema;
-    }
-    return readerContext.getSchemaFromBufferRecord(bufferedRecord);
+    return bufferedRecordMerger.finalMerge(olderRecord, newerRecord);
   }
 
   protected boolean hasNextBaseRecord(T baseRecord, BufferedRecord<T> logRecordInfo) throws IOException {
@@ -638,19 +389,5 @@ public abstract class FileGroupRecordBuffer<T> implements HoodieFileGroupRecordB
     return isCommitTimeOrderingValue(deleteRecord.getOrderingValue())
         ? DEFAULT_ORDERING_VALUE
         : readerContext.convertValueToEngineType(deleteRecord.getOrderingValue());
-  }
-
-  private boolean isDeleteRecord(Option<T> record, Schema schema) {
-    if (record.isEmpty()) {
-      return true;
-    }
-
-    Object operation = readerContext.getValue(record.get(), schema, OPERATION_METADATA_FIELD);
-    if (operation != null && HoodieOperation.isDeleteRecord(operation.toString())) {
-      return true;
-    }
-
-    Object deleteMarker = readerContext.getValue(record.get(), schema, HOODIE_IS_DELETED_FIELD);
-    return deleteMarker instanceof Boolean && (boolean) deleteMarker;
   }
 }
